Added about me with subscribed? boolean
Added about me function that return about me page and a boolean indicating if the logged-in user is subscribed to the target analyst.

Added subscription check method to subscriptionController.java that can be used to check if a user is subscribed to a target user.

Also added Javadoc to DTO classes that were missing.