Refactored code to use new NoSuchUserException
Code now uses the custom NoSuchUserException when a user can't be found. The old exception was not meant to be used for more than login handling. This meant that the exception was handled by Spring automatically by trying to redirect the user to the login page.

Also allowed the /error mapping which was previously blocked.